[refactor] parecer: diretorios front #2150
