+++ front/src/modules/parecer/components/gerenciar-paineis/GerenciarListaAguardandoDistribuicao.vue
@@ -54,7 +54,7 @@
 <script>
 
 import MxUtils from '@/mixins/utils';
-import MxUtilsParecer from '@/modules/parecer/mixins/utilsParecer';
+import MxUtilsParecer from '@/modules/parecer/mixins/UtilsParecer';
 import TdNomeProduto from '@/modules/parecer/components/gerenciar-paineis/TdNomeProduto';
 import TdTipoProduto from '@/modules/parecer/components/gerenciar-paineis/TdTipoProduto';
 import TdNumeroPronac from '@/modules/parecer/components/gerenciar-paineis/TdNumeroPronac';
