+++ front/src/modules/parecer/components/gerenciar-paineis/GerenciarListaDevolvidasSecult.vue
@@ -164,7 +164,7 @@
 <script>
 
 import MxUtils from '@/mixins/utils';
-import MxUtilsParecer from '@/modules/parecer/mixins/utilsParecer';
+import MxUtilsParecer from '@/modules/parecer/mixins/UtilsParecer';
 
 import SalicTextoSimples from '@/components/SalicTextoSimples';
 import TdNomeProduto from '@/modules/parecer/components/gerenciar-paineis/TdNomeProduto';
