+++ front/src/modules/parecer/mixins/UtilsParecer.js
